refactor: move src/consumer/component/dependencies-resolver.js to src/consumer/component/dependencies/dependency-resolver/dependencies-resolver.js, also, re-use the types from dependency-resolver/types component
