+++ src/consumer/component/dependencies/dependency.js
@@ -1,26 +1,7 @@
 /** @flow */
 import { BitId } from '../../../bit-id';
 import type { PathLinux } from '../../../utils/path';
-
-/**
- * Import Specifier data.
- * For example, `import foo from './bar' `, "foo" is the import-specifier and is default.
- * Conversely, `import { foo } from './bar' `, here, "foo" is non-default.
- */
-type Specifier = {
-  isDefault: boolean,
-  name: string
-};
-
-/**
- * ImportSpecifier are used to generate links from component to its dependencies.
- * For example, a component might have a dependency: "import { foo } from './bar' ", when a link is generated, we use
- * the import-specifier name, which is "foo" to generate the link correctly.
- */
-export type ImportSpecifier = {
-  mainFile: Specifier,
-  linkFile?: Specifier // relevant only when the dependency is a link file (e.g. index.js which import and export the variable from other file)
-};
+import type { ImportSpecifier } from './dependency-resolver/types/dependency-tree-type';
 
 export type RelativePath = {
   sourceRelativePath: PathLinux,
